+++ packages/experimental/Text/src/Text.tsx
@@ -77,7 +77,7 @@ export const Text = compressible<TextProps, TextTokens>((props: TextProps, useTo
 
   // return a continuation function that allows this text to be compressed
   return (extra: TextProps, children: React.ReactNode) => {
-    const mergedProps = { ...rest, ...extra, style: mergeStyles(tokenStyle, props.style, extra.style) };
+    const mergedProps = { ...rest, ...extra, style: mergeStyles(tokenStyle, props.style, extra?.style) };
     return <RNText {...mergedProps}>{children}</RNText>;
   };
 }, useTextTokens);