+++ packages/components/Persona/CHANGELOG.md
@@ -1,8 +1,39 @@
 # Change Log - @fluentui-react-native/persona
 
-This log was last generated on Fri, 18 Feb 2022 23:27:11 GMT and should not be manually modified.
+This log was last generated on Mon, 07 Mar 2022 19:15:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.17
+
+Mon, 07 Mar 2022 19:15:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/adapters to v0.8.4
+- Bump @fluentui-react-native/framework to v0.7.17
+- Bump @fluentui-react-native/persona-coin to v0.10.17
+- Bump @fluentui-react-native/tokens to v0.11.8
+
+## 0.12.16
+
+Fri, 04 Mar 2022 23:43:10 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.7.16
+- Bump @fluentui-react-native/persona-coin to v0.10.16
+
+## 0.12.15
+
+Thu, 03 Mar 2022 20:20:09 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.11.11
+- Bump @fluentui-react-native/framework to v0.7.15
+- Bump @fluentui-react-native/persona-coin to v0.10.15
+- Bump @fluentui-react-native/tokens to v0.11.7
 
 ## 0.12.14