+++ packages/theming/theme-types/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/theme-types
 
-This log was last generated on Fri, 14 Jan 2022 01:00:04 GMT and should not be manually modified.
+This log was last generated on Thu, 03 Mar 2022 20:20:09 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.0
+
+Thu, 03 Mar 2022 20:20:09 GMT
+
+### Minor changes
+
+- New alias tokens (ruaraki@microsoft.com)
 
 ## 0.13.3