+++ packages/utils/adapters/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/adapters
 
-This log was last generated on Fri, 18 Feb 2022 23:27:10 GMT and should not be manually modified.
+This log was last generated on Mon, 07 Mar 2022 19:15:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.4
+
+Mon, 07 Mar 2022 19:15:33 GMT
+
+### Patches
+
+- Pass tooltip to React Native macOS components (sanajmi@microsoft.com)
 
 ## 0.8.3