+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Tue, 01 Mar 2022 20:18:31 GMT and should not be manually modified.
+This log was last generated on Mon, 07 Mar 2022 19:15:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.3
+
+Mon, 07 Mar 2022 19:15:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.12
+
+## 0.14.2
+
+Fri, 04 Mar 2022 23:43:10 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.11
+
+## 0.14.1
+
+Thu, 03 Mar 2022 20:20:09 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.10
 
 ## 0.14.0